improve Claude extraction: tool use, caching, retries, token limit, date removal
- Switch from free-text JSON to tool_use for deterministic schema-enforced output
- Add prompt caching (anthropic-beta) on system prompt + tool definition
- Raise default max_tokens 2048->4096; warn in logs when limit is hit
- Add configurable retries (Claude:MaxRetries=3) with randomised jitter delays
- Remove dateOfQuote/estimatedDeliveryDate from extraction; dates come from RFQ record
- Warn+note in content when email body is truncated at Claude:MaxContentChars
- Add injection defence note to system prompt
- ProductLine.UnitsRequested/UnitsQuoted: int? -> double? for fractional quantities
- SharePointService: extract shared Str/GetStr/RfqId/IsAppField/ParentFields helpers,
  remove duplicated local functions across ReadSupplierItemsAsync,
  ReadSupplierItemsByRfqIdAsync, GetNewResponsesSinceAsync, GetNewRfqReferencesSinceAsync
- appsettings.json: document MaxTokens, add MaxRetries key

+++ OutlookShredder.Proxy/Models/ProductLine.cs
@@ -3,8 +3,8 @@ namespace OutlookShredder.Proxy.Models;
 public class ProductLine
 {
     public string? ProductName             { get; set; }
-    public int?    UnitsRequested          { get; set; }
-    public int?    UnitsQuoted             { get; set; }
+    public double? UnitsRequested          { get; set; }
+    public double? UnitsQuoted             { get; set; }
     public double? LengthPerUnit           { get; set; }
     public string? LengthUnit              { get; set; }
     public double? WeightPerUnit           { get; set; }
